--- auto_machinelearning/ml/regression/model_params.yaml
@@ -1,29 +1,0 @@
-Linear Regression: {}
-
-Decision Tree:
-  max_depth: [null, 5, 10]
-  min_samples_split: [2, 5, 10, 20]
-  min_samples_leaf: [1, 5, 10, 20]
-  max_features: ["sqrt", "log2"]
-  criterion: ["squared_error", "friedman_mse"]
-
-Random Forest:
-  n_estimators: [100, 200]
-  max_depth: [null, 5, 10]
-  min_samples_split: [2, 5]
-  min_samples_leaf: [1, 2, 4]
-  max_features: ["sqrt", "log2"]
-  bootstrap: [true, false]
-
-Gradiant Boosting:
-  n_estimators: [100, 200]
-  learning_rate: [0.01, 0.1]
-  max_depth: [3, 5]
-  min_samples_split: [2, 5]
-  min_samples_leaf: [1, 3]
-  max_features: ["sqrt", "log2"]
-  subsample: [0.7, 0.8]
-
-Ridge Regression:
-  alpha: [0.001, 0.01]
-
