+++ ml_genius/ml/classification/components/model_trainer.py
@@ -16,7 +16,7 @@ from sklearn.ensemble import AdaBoostClassifier
 from sklearn.model_selection import GridSearchCV, RandomizedSearchCV
 from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
 
-from auto_machinelearning.ml.classification.entity.artifacts_entity import ClassificationMetricArtifact, ModelTrainerArtifact, DataTransformationArtifacts
+from ml_genius.ml.classification.entity.artifacts_entity import ClassificationMetricArtifact, ModelTrainerArtifact, DataTransformationArtifacts
 
 
 class ModelTrainer:
